pyproject.toml
[tool.poetry]
name = "advancedrag"
version = "0.1.0"
description = ""
authors = ["AbdullahAlzariqi <abdallahtalat5597@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.9"
langgraph = "^0.2.56"
langchain-openai = "^0.2.11"
langchain-google-genai = "^2.0.6"
black = "^24.10.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
langchainhub = "^0.1.21"
langchain-community = "^0.3.9"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
